PIM-3895: rename ValueSetter to AttributeSetter, rework ProductUpdater and related specs
