+++ spec/Pim/Bundle/CatalogBundle/Updater/ProductUpdaterSpec.php
@@ -12,6 +12,8 @@ use Pim\Bundle\CatalogBundle\Updater\Adder\AttributeAdderInterface;
 use Pim\Bundle\CatalogBundle\Updater\Adder\FieldAdderInterface;
 use Pim\Bundle\CatalogBundle\Updater\Copier\CopierInterface;
 use Pim\Bundle\CatalogBundle\Updater\Copier\CopierRegistryInterface;
+use Pim\Bundle\CatalogBundle\Updater\Setter\AttributeSetterInterface;
+use Pim\Bundle\CatalogBundle\Updater\Setter\FieldSetterInterface;
 use Pim\Bundle\CatalogBundle\Updater\Setter\SetterInterface;
 use Pim\Bundle\CatalogBundle\Updater\Setter\SetterRegistryInterface;
 use Prophecy\Argument;
@@ -38,13 +40,14 @@ class ProductUpdaterSpec extends ObjectBehavior
         ProductInterface $product1,
         ProductInterface $product2,
         AttributeInterface $attribute,
-        SetterInterface $setter
+        AttributeSetterInterface $setter
     ) {
         $products = [$product1, $product2];
 
         $attributeRepository->findOneBy(['code' => 'field'])->willReturn($attribute);
-        $setterRegistry->get($attribute)->willReturn($setter);
-        $setter->setValue($products, $attribute, 'data', 'fr_FR', 'ecommerce')->shouldBeCalled();
+        $setterRegistry->getAttributeSetter($attribute)->willReturn($setter);
+        $setter->setAttributeData($product1, $attribute, 'data', ['locale' => 'fr_FR', 'scope' => 'ecommerce'])->shouldBeCalled();
+        $setter->setAttributeData($product2, $attribute, 'data', ['locale' => 'fr_FR', 'scope' => 'ecommerce'])->shouldBeCalled();
 
         $this->setValue($products, 'field', 'data', 'fr_FR', 'ecommerce');
     }
@@ -52,7 +55,7 @@ class ProductUpdaterSpec extends ObjectBehavior
     function it_throws_an_exception_when_it_sets_an_unknown_field($attributeRepository, ProductInterface $product)
     {
         $attributeRepository->findOneBy(Argument::any())->willReturn(null);
-        $this->shouldThrow(new \LogicException('Unknown attribute "unknown_field".'))->during(
+        $this->shouldThrow(new \LogicException('No setter found for field "unknown_field"'))->during(
             'setValue', [[$product], 'unknown_field', 'data', 'fr_FR', 'ecommerce']
         );
     }
@@ -78,6 +81,45 @@ class ProductUpdaterSpec extends ObjectBehavior
         $this->copyValue($products, 'from_field', 'to_field', 'from_locale', 'to_locale', 'from_scope', 'to_scope');
     }
 
+    function it_throws_an_exception_when_it_copies_an_unknown_field($attributeRepository, ProductInterface $product)
+    {
+        $attributeRepository->findOneBy(Argument::any())->willReturn(null);
+        $this->shouldThrow(new \LogicException('Unknown attribute "unknown_field".'))->during(
+            'copyValue', [[$product], 'unknown_field', 'to_field', 'from_locale', 'to_locale', 'from_scope', 'to_scope']
+        );
+    }
+
+    function it_sets_a_data_to_a_product_attribute(
+        $setterRegistry,
+        $attributeRepository,
+        ProductInterface $product,
+        AttributeInterface $attribute,
+        AttributeSetterInterface $setter
+    ) {
+        $attributeRepository->findOneBy(['code' => 'name'])->willReturn($attribute);
+        $setterRegistry->getAttributeSetter($attribute)->willReturn($setter);
+        $setter
+            ->setAttributeData($product, $attribute, 'my name', [])
+            ->shouldBeCalled();
+
+        $this->setData($product, 'name', 'my name', []);
+    }
+
+    function it_sets_a_data_to_a_product_field(
+        $setterRegistry,
+        $attributeRepository,
+        ProductInterface $product,
+        FieldSetterInterface $setter
+    ) {
+        $attributeRepository->findOneBy(['code' => 'category'])->willReturn(null);
+        $setterRegistry->getFieldSetter('category')->willReturn($setter);
+        $setter
+            ->setFieldData($product, 'category', ['tshirt'], [])
+            ->shouldBeCalled();
+
+        $this->setData($product, 'category', ['tshirt'], []);
+    }
+
     function it_adds_a_data_to_a_product_attribute(
         $adderRegistry,
         $attributeRepository,
@@ -85,13 +127,13 @@ class ProductUpdaterSpec extends ObjectBehavior
         AttributeInterface $attribute,
         AttributeAdderInterface $adder
     ) {
-        $attributeRepository->findOneBy(['code' => 'name'])->willReturn($attribute);
+        $attributeRepository->findOneBy(['code' => 'color'])->willReturn($attribute);
         $adderRegistry->getAttributeAdder($attribute)->willReturn($adder);
         $adder
-            ->addAttributeData($product, $attribute, 'my name', [])
+            ->addAttributeData($product, $attribute, ['red', 'blue'], [])
             ->shouldBeCalled();
 
-        $this->addData($product, 'name', 'my name', []);
+        $this->addData($product, 'color', ['red', 'blue'], []);
     }
 
     function it_adds_a_data_to_a_product_field(
@@ -107,13 +149,5 @@ class ProductUpdaterSpec extends ObjectBehavior
             ->shouldBeCalled();
 
         $this->addData($product, 'category', 'tshirt', []);
-    }
-
-    function it_throws_an_exception_when_it_copies_an_unknown_field($attributeRepository, ProductInterface $product)
-    {
-        $attributeRepository->findOneBy(Argument::any())->willReturn(null);
-        $this->shouldThrow(new \LogicException('Unknown attribute "unknown_field".'))->during(
-            'copyValue', [[$product], 'unknown_field', 'to_field', 'from_locale', 'to_locale', 'from_scope', 'to_scope']
-        );
     }
 }
