+++ spec/Pim/Bundle/CatalogBundle/Updater/Setter/BooleanAttributeSetterSpec.php
@@ -12,11 +12,11 @@ use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 use Prophecy\Argument;
 
-class TextValueSetterSpec extends ObjectBehavior
+class BooleanAttributeSetterSpec extends ObjectBehavior
 {
     function let(ProductBuilderInterface $builder, AttributeValidatorHelper $attrValidatorHelper)
     {
-        $this->beConstructedWith($builder, $attrValidatorHelper, ['pim_catalog_text', 'pim_catalog_textarea']);
+        $this->beConstructedWith($builder, $attrValidatorHelper, ['pim_catalog_boolean']);
     }
 
     function it_is_a_setter()
@@ -24,83 +24,78 @@ class TextValueSetterSpec extends ObjectBehavior
         $this->shouldImplement('Pim\Bundle\CatalogBundle\Updater\Setter\SetterInterface');
     }
 
-    function it_supports_text_attributes(
-        AttributeInterface $textAttribute,
-        AttributeInterface $textareaAttribute,
-        AttributeInterface $numberAttribute
+    function it_supports_boolean_attributes(
+        AttributeInterface $booleanAttribute,
+        AttributeInterface $textareaAttribute
     ) {
-        $textAttribute->getAttributeType()->willReturn('pim_catalog_text');
-        $this->supports($textAttribute)->shouldReturn(true);
-        $this->supportsAttribute($textAttribute)->shouldReturn(true);
+        $booleanAttribute->getAttributeType()->willReturn('pim_catalog_boolean');
+        $this->supports($booleanAttribute)->shouldReturn(true);
+        $this->supportsAttribute($booleanAttribute)->shouldReturn(true);
 
         $textareaAttribute->getAttributeType()->willReturn('pim_catalog_textarea');
-        $this->supports($textareaAttribute)->shouldReturn(true);
-        $this->supportsAttribute($textareaAttribute)->shouldReturn(true);
-
-        $numberAttribute->getAttributeType()->willReturn('pim_catalog_number');
-        $this->supports($numberAttribute)->shouldReturn(false);
-        $this->supportsAttribute($numberAttribute)->shouldReturn(false);
+        $this->supports($textareaAttribute)->shouldReturn(false);
+        $this->supportsAttribute($textareaAttribute)->shouldReturn(false);
     }
 
     function it_checks_locale_and_scope_when_setting_a_value(
         $attrValidatorHelper,
         AttributeInterface $attribute,
         ProductInterface $product,
-        ProductValueInterface $textValue
+        ProductValueInterface $booleanValue
     ) {
         $attrValidatorHelper->validateLocale(Argument::cetera())->shouldBeCalled();
         $attrValidatorHelper->validateScope(Argument::cetera())->shouldBeCalled();
 
-        $attribute->getCode()->willReturn('description');
-        $product->getValue('description', 'fr_FR', 'mobile')->willReturn($textValue);
-        $textValue->setData('data');
+        $attribute->getCode()->willReturn('displayable');
+        $product->getValue('displayable', 'fr_FR', 'mobile')->willReturn($booleanValue);
+        $booleanValue->setData(true)->shouldBeCalled();
 
-        $this->setValue([$product], $attribute, 'data', 'fr_FR', 'mobile');
+        $this->setValue([$product], $attribute, true, 'fr_FR', 'mobile');
     }
 
     function it_checks_locale_and_scope_when_setting_an_attribute_data(
         $attrValidatorHelper,
         AttributeInterface $attribute,
         ProductInterface $product,
-        ProductValueInterface $textValue
+        ProductValueInterface $booleanValue
     ) {
         $attrValidatorHelper->validateLocale(Argument::cetera())->shouldBeCalled();
         $attrValidatorHelper->validateScope(Argument::cetera())->shouldBeCalled();
 
-        $attribute->getCode()->willReturn('description');
-        $product->getValue('description', 'fr_FR', 'mobile')->willReturn($textValue);
-        $textValue->setData('data');
+        $attribute->getCode()->willReturn('displayable');
+        $product->getValue('displayable', 'fr_FR', 'mobile')->willReturn($booleanValue);
+        $booleanValue->setData(true)->shouldBeCalled();
 
-        $this->setAttributeData($product, $attribute, 'data', ['locale' => 'fr_FR', 'scope' => 'mobile']);
+        $this->setAttributeData($product, $attribute, true, ['locale' => 'fr_FR', 'scope' => 'mobile']);
     }
 
-    function it_throws_an_error_if_data_is_not_a_string(
+    function it_throws_an_error_if_data_is_not_a_boolean(
         AttributeInterface $attribute,
         ProductInterface $product
     ) {
         $attribute->getCode()->willReturn('attributeCode');
 
-        $data = 42;
+        $data = 'not a boolean';
 
         $this->shouldThrow(
-            InvalidArgumentException::stringExpected('attributeCode', 'setter', 'text', gettype($data))
+            InvalidArgumentException::booleanExpected('attributeCode', 'setter', 'boolean', gettype($data))
         )->during('setValue', [[$product], $attribute, $data, 'fr_FR', 'mobile']);
     }
 
-    function it_throws_an_error_if_attribute_data_is_not_a_string(
+    function it_throws_an_error_if_attribute_data_is_not_a_boolean(
         AttributeInterface $attribute,
         ProductInterface $product
     ) {
         $attribute->getCode()->willReturn('attributeCode');
 
-        $data = 42;
+        $data = 'not a boolean';
 
         $this->shouldThrow(
-            InvalidArgumentException::stringExpected('attributeCode', 'setter', 'text', gettype($data))
+            InvalidArgumentException::booleanExpected('attributeCode', 'setter', 'boolean', gettype($data))
         )->during('setAttributeData', [$product, $attribute, $data, ['locale' => 'fr_FR', 'scope' => 'mobile']]);
     }
 
-    function it_sets_text_value_to_a_product_value(
+    function it_sets_boolean_value_to_a_product_value(
         AttributeInterface $attribute,
         ProductInterface $product1,
         ProductInterface $product2,
@@ -110,7 +105,7 @@ class TextValueSetterSpec extends ObjectBehavior
     ) {
         $locale = 'fr_FR';
         $scope = 'mobile';
-        $data = 'data';
+        $data = true;
 
         $attribute->getCode()->willReturn('attributeCode');
         $productValue->setData($data)->shouldBeCalled();
@@ -128,7 +123,7 @@ class TextValueSetterSpec extends ObjectBehavior
         $this->setValue($products, $attribute, $data, $locale, $scope);
     }
 
-    function it_sets_attribute_data_text_value_to_a_product_value(
+    function it_sets_attribute_data_boolean_value_to_a_product_value(
         AttributeInterface $attribute,
         ProductInterface $product1,
         ProductInterface $product2,
@@ -138,7 +133,7 @@ class TextValueSetterSpec extends ObjectBehavior
     ) {
         $locale = 'fr_FR';
         $scope = 'mobile';
-        $data = 'data';
+        $data = true;
 
         $attribute->getCode()->willReturn('attributeCode');
         $productValue->setData($data)->shouldBeCalled();
