+++ spec/Pim/Bundle/CatalogBundle/Updater/Setter/DateAttributeSetterSpec.php
@@ -11,7 +11,7 @@ use Pim\Bundle\CatalogBundle\Exception\InvalidArgumentException;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 use Prophecy\Argument;
 
-class DateValueSetterSpec extends ObjectBehavior
+class DateAttributeSetterSpec extends ObjectBehavior
 {
     function let(ProductBuilderInterface $builder, AttributeValidatorHelper $attrValidatorHelper)
     {
