+++ spec/Pim/Bundle/CatalogBundle/Updater/Setter/MetricAttributeSetterSpec.php
@@ -15,7 +15,7 @@ use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 use Prophecy\Argument;
 
-class MetricValueSetterSpec extends ObjectBehavior
+class MetricAttributeSetterSpec extends ObjectBehavior
 {
     function let(
         ProductBuilderInterface $builder,
