+++ spec/Pim/Bundle/CatalogBundle/Updater/Setter/MultiSelectAttributeSetterSpec.php
@@ -13,7 +13,7 @@ use Pim\Bundle\CatalogBundle\Repository\AttributeOptionRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 use Prophecy\Argument;
 
-class MultiSelectValueSetterSpec extends ObjectBehavior
+class MultiSelectAttributeSetterSpec extends ObjectBehavior
 {
     function let(
         ProductBuilderInterface $builder,
