+++ spec/Pim/Bundle/CatalogBundle/Updater/Setter/PriceCollectionAttributeSetterSpec.php
@@ -14,7 +14,7 @@ use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 use Prophecy\Argument;
 
-class PriceCollectionValueSetterSpec extends ObjectBehavior
+class PriceCollectionAttributeSetterSpec extends ObjectBehavior
 {
     function let(
         ProductBuilderInterface $builder,
