+++ spec/Pim/Bundle/CatalogBundle/Updater/Setter/TextAttributeSetterSpec.php
@@ -12,11 +12,11 @@ use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 use Prophecy\Argument;
 
-class NumberValueSetterSpec extends ObjectBehavior
+class TextAttributeSetterSpec extends ObjectBehavior
 {
     function let(ProductBuilderInterface $builder, AttributeValidatorHelper $attrValidatorHelper)
     {
-        $this->beConstructedWith($builder, $attrValidatorHelper, ['pim_catalog_number']);
+        $this->beConstructedWith($builder, $attrValidatorHelper, ['pim_catalog_text', 'pim_catalog_textarea']);
     }
 
     function it_is_a_setter()
@@ -24,78 +24,83 @@ class NumberValueSetterSpec extends ObjectBehavior
         $this->shouldImplement('Pim\Bundle\CatalogBundle\Updater\Setter\SetterInterface');
     }
 
-    function it_supports_number_attributes(
-        AttributeInterface $numberAttribute,
-        AttributeInterface $textareaAttribute
+    function it_supports_text_attributes(
+        AttributeInterface $textAttribute,
+        AttributeInterface $textareaAttribute,
+        AttributeInterface $numberAttribute
     ) {
-        $numberAttribute->getAttributeType()->willReturn('pim_catalog_number');
-        $this->supports($numberAttribute)->shouldReturn(true);
-        $this->supportsAttribute($numberAttribute)->shouldReturn(true);
+        $textAttribute->getAttributeType()->willReturn('pim_catalog_text');
+        $this->supports($textAttribute)->shouldReturn(true);
+        $this->supportsAttribute($textAttribute)->shouldReturn(true);
 
         $textareaAttribute->getAttributeType()->willReturn('pim_catalog_textarea');
-        $this->supports($textareaAttribute)->shouldReturn(false);
-        $this->supportsAttribute($textareaAttribute)->shouldReturn(false);
+        $this->supports($textareaAttribute)->shouldReturn(true);
+        $this->supportsAttribute($textareaAttribute)->shouldReturn(true);
+
+        $numberAttribute->getAttributeType()->willReturn('pim_catalog_number');
+        $this->supports($numberAttribute)->shouldReturn(false);
+        $this->supportsAttribute($numberAttribute)->shouldReturn(false);
     }
 
     function it_checks_locale_and_scope_when_setting_a_value(
         $attrValidatorHelper,
         AttributeInterface $attribute,
         ProductInterface $product,
-        ProductValueInterface $numberValue
+        ProductValueInterface $textValue
     ) {
         $attrValidatorHelper->validateLocale(Argument::cetera())->shouldBeCalled();
         $attrValidatorHelper->validateScope(Argument::cetera())->shouldBeCalled();
 
-        $attribute->getCode()->willReturn('response_to_universe');
-        $product->getValue('response_to_universe', 'fr_FR', 'mobile')->willReturn($numberValue);
-        $numberValue->setData(42)->shouldBeCalled();
+        $attribute->getCode()->willReturn('description');
+        $product->getValue('description', 'fr_FR', 'mobile')->willReturn($textValue);
+        $textValue->setData('data');
 
-        $this->setValue([$product], $attribute, 42, 'fr_FR', 'mobile');
+        $this->setValue([$product], $attribute, 'data', 'fr_FR', 'mobile');
     }
 
     function it_checks_locale_and_scope_when_setting_an_attribute_data(
         $attrValidatorHelper,
         AttributeInterface $attribute,
         ProductInterface $product,
-        ProductValueInterface $numberValue
+        ProductValueInterface $textValue
     ) {
         $attrValidatorHelper->validateLocale(Argument::cetera())->shouldBeCalled();
         $attrValidatorHelper->validateScope(Argument::cetera())->shouldBeCalled();
 
-        $attribute->getCode()->willReturn('response_to_universe');
-        $product->getValue('response_to_universe', 'fr_FR', 'mobile')->willReturn($numberValue);
-        $numberValue->setData(42)->shouldBeCalled();
+        $attribute->getCode()->willReturn('description');
+        $product->getValue('description', 'fr_FR', 'mobile')->willReturn($textValue);
+        $textValue->setData('data');
 
-        $this->setAttributeData($product, $attribute, 42, ['locale' => 'fr_FR', 'scope' => 'mobile']);
+        $this->setAttributeData($product, $attribute, 'data', ['locale' => 'fr_FR', 'scope' => 'mobile']);
     }
 
-    function it_throws_an_error_if_data_is_not_a_number_or_null(
+    function it_throws_an_error_if_data_is_not_a_string(
         AttributeInterface $attribute,
         ProductInterface $product
     ) {
         $attribute->getCode()->willReturn('attributeCode');
 
-        $data = 'not a number';
+        $data = 42;
 
         $this->shouldThrow(
-            InvalidArgumentException::numericExpected('attributeCode', 'setter', 'number', gettype($data))
+            InvalidArgumentException::stringExpected('attributeCode', 'setter', 'text', gettype($data))
         )->during('setValue', [[$product], $attribute, $data, 'fr_FR', 'mobile']);
     }
 
-    function it_throws_an_error_if_attribute_data_is_not_a_number_or_null(
+    function it_throws_an_error_if_attribute_data_is_not_a_string(
         AttributeInterface $attribute,
         ProductInterface $product
     ) {
         $attribute->getCode()->willReturn('attributeCode');
 
-        $data = 'not a number';
+        $data = 42;
 
         $this->shouldThrow(
-            InvalidArgumentException::numericExpected('attributeCode', 'setter', 'number', gettype($data))
+            InvalidArgumentException::stringExpected('attributeCode', 'setter', 'text', gettype($data))
         )->during('setAttributeData', [$product, $attribute, $data, ['locale' => 'fr_FR', 'scope' => 'mobile']]);
     }
 
-    function it_sets_number_value_to_a_product_value(
+    function it_sets_text_value_to_a_product_value(
         AttributeInterface $attribute,
         ProductInterface $product1,
         ProductInterface $product2,
@@ -105,7 +110,7 @@ class NumberValueSetterSpec extends ObjectBehavior
     ) {
         $locale = 'fr_FR';
         $scope = 'mobile';
-        $data = 44;
+        $data = 'data';
 
         $attribute->getCode()->willReturn('attributeCode');
         $productValue->setData($data)->shouldBeCalled();
@@ -123,7 +128,7 @@ class NumberValueSetterSpec extends ObjectBehavior
         $this->setValue($products, $attribute, $data, $locale, $scope);
     }
 
-    function it_sets_number_value_to_a_product_attribute_data_value(
+    function it_sets_attribute_data_text_value_to_a_product_value(
         AttributeInterface $attribute,
         ProductInterface $product1,
         ProductInterface $product2,
@@ -133,7 +138,7 @@ class NumberValueSetterSpec extends ObjectBehavior
     ) {
         $locale = 'fr_FR';
         $scope = 'mobile';
-        $data = 44;
+        $data = 'data';
 
         $attribute->getCode()->willReturn('attributeCode');
         $productValue->setData($data)->shouldBeCalled();
@@ -145,8 +150,6 @@ class NumberValueSetterSpec extends ObjectBehavior
         $product1->getValue('attributeCode', $locale, $scope)->shouldBeCalled()->willReturn($productValue);
         $product2->getValue('attributeCode', $locale, $scope)->willReturn(null);
         $product3->getValue('attributeCode', $locale, $scope)->willReturn($productValue);
-
-        $products = [$product1, $product2, $product3];
 
         $this->setAttributeData($product1, $attribute, $data, ['locale' => $locale, 'scope' => $scope]);
         $this->setAttributeData($product2, $attribute, $data, ['locale' => $locale, 'scope' => $scope]);
