+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/DateAttributeSetter.php
@@ -4,19 +4,18 @@ namespace Pim\Bundle\CatalogBundle\Updater\Setter;
 
 use Pim\Bundle\CatalogBundle\Builder\ProductBuilderInterface;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
-use Pim\Bundle\CatalogBundle\Builder\ProductBuilder;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\CatalogBundle\Exception\InvalidArgumentException;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 
 /**
- * Sets a number value in many products
+ * Sets a date value in many products
  *
  * @author    Olivier Soulet <olivier.soulet@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class NumberValueSetter extends AbstractAttributeSetter
+class DateAttributeSetter extends AbstractAttributeSetter
 {
     /**
      * @param ProductBuilderInterface  $productBuilder
@@ -47,7 +46,7 @@ class NumberValueSetter extends AbstractAttributeSetter
     /**
      * {@inheritdoc}
      *
-     * Expected data input format: "12.0"|"12"|12|12.3
+     * Expected data input format : "yyyy-mm-dd"
      */
     public function setAttributeData(
         ProductInterface $product,
@@ -56,14 +55,42 @@ class NumberValueSetter extends AbstractAttributeSetter
         array $options = []
     ) {
         $this->resolver->resolve($options);
-        $this->checkLocaleAndScope($attribute, $options['locale'], $options['scope'], 'number');
-        $this->checkData($attribute, $data);
+        $this->checkLocaleAndScope($attribute, $options['locale'], $options['scope'], 'date');
+        $data = $this->formatData($attribute, $data);
 
         $this->setData($product, $attribute, $data, $options['locale'], $options['scope']);
     }
 
     /**
-     * Set data into product value
+     * Format data
+     *
+     * @param AttributeInterface $attribute
+     * @param mixed              $data
+     *
+     * @return string
+     */
+    protected function formatData(AttributeInterface $attribute, $data)
+    {
+        if ($data instanceof \DateTime) {
+            $data = $data->format('Y-m-d');
+        } elseif (is_string($data)) {
+            $this->validateDateFormat($attribute, $data);
+        } elseif (null !== $data) {
+            throw InvalidArgumentException::expected(
+                $attribute->getCode(),
+                'datetime or string',
+                gettype($data),
+                'setter',
+                'date',
+                $data
+            );
+        }
+
+        return $data;
+    }
+
+    /**
+     * Set the data into the product value
      *
      * @param ProductInterface   $product
      * @param AttributeInterface $attribute
@@ -77,18 +104,35 @@ class NumberValueSetter extends AbstractAttributeSetter
         if (null === $value) {
             $value = $this->productBuilder->addProductValue($product, $attribute, $locale, $scope);
         }
+
+        if (null !== $data) {
+            $data = new \DateTime($data);
+        }
+
         $value->setData($data);
     }
 
     /**
-     * Check data input
      * @param AttributeInterface $attribute
-     * @param mixed              $data
+     * @param string             $data
      */
-    protected function checkData(AttributeInterface $attribute, $data)
+    protected function validateDateFormat(AttributeInterface $attribute, $data)
     {
-        if (!is_numeric($data) && null !== $data) {
-            throw InvalidArgumentException::numericExpected($attribute->getCode(), 'setter', 'number', gettype($data));
+        $dateValues = explode('-', $data);
+
+        if (
+            count($dateValues) !== 3
+            || (!is_numeric($dateValues[0]) || !is_numeric($dateValues[1]) || !is_numeric($dateValues[2]))
+            || !checkdate($dateValues[1], $dateValues[2], $dateValues[0])
+        ) {
+            throw InvalidArgumentException::expected(
+                $attribute->getCode(),
+                'a string with the format yyyy-mm-dd',
+                'setter',
+                'date',
+                gettype($data),
+                $data
+            );
         }
     }
 }
