+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/MediaAttributeSetter.php
@@ -19,7 +19,7 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class MediaValueSetter extends AbstractAttributeSetter
+class MediaAttributeSetter extends AbstractAttributeSetter
 {
     /** @var MediaManager */
     protected $mediaManager;
