+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/MultiSelectAttributeSetter.php
@@ -16,7 +16,7 @@ use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class MultiSelectValueSetter extends AbstractAttributeSetter
+class MultiSelectAttributeSetter extends AbstractAttributeSetter
 {
     /** @var AttributeOptionRepositoryInterface */
     protected $attrOptionRepository;
