+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/PriceCollectionAttributeSetter.php
@@ -16,7 +16,7 @@ use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class PriceCollectionValueSetter extends AbstractAttributeSetter
+class PriceCollectionAttributeSetter extends AbstractAttributeSetter
 {
     /** @var CurrencyManager */
     protected $currencyManager;
