+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/SimpleSelectAttributeSetter.php
@@ -17,7 +17,7 @@ use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class SimpleSelectValueSetter extends AbstractAttributeSetter
+class SimpleSelectAttributeSetter extends AbstractAttributeSetter
 {
     /** @var AttributeOptionRepositoryInterface */
     protected $attrOptionRepository;
