+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/TextAttributeSetter.php
@@ -15,7 +15,7 @@ use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class TextValueSetter extends AbstractAttributeSetter
+class TextAttributeSetter extends AbstractAttributeSetter
 {
     /**
      * @param ProductBuilderInterface  $productBuilder
